+++ lib/uploadcare/param/authentication_header.rb
@@ -11,6 +11,7 @@ module Uploadcare
     class AuthenticationHeader
       # @see https://uploadcare.com/docs/api_reference/rest/requests_auth/#auth-uploadcare
       def self.call(options = {})
+        validate_auth_config
         case Uploadcare.config.auth_type
         when 'Uploadcare'
           SecureAuthHeader.call(options)
@@ -20,6 +21,17 @@ module Uploadcare
           raise ArgumentError, "Unknown auth_scheme: '#{Uploadcare.config.auth_type}'"
         end
       end
+
+      def self.validate_auth_config
+        raise AuthError, 'Public Key is blank.' if is_blank?(Uploadcare.config.public_key)
+        raise AuthError, 'Secret Key is blank.' if is_blank?(Uploadcare.config.secret_key)
+      end
+
+      # rubocop:disable Naming/PredicateName
+      def self.is_blank?(value)
+        value.nil? || value.empty?
+      end
+      # rubocop:enable Naming/PredicateName
     end
   end
 end