+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 ### Changed
 * File attribute `datetime_stored` is deprecated and will warn on usage from `File` object properties.
 
+## Unreleased
+
+### Fixed
+
+* Throw `AuthError` if current public key or secret key config are empty when accessing any of the APIs. 
+
 ## 4.4.0 — 2024-03-09
 
 ### Breaking